nlp-service/requirements.txt
# NLP Service Dependencies
# Python FastAPI microservice for resume parsing and skill extraction

# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# NLP & ML
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2

# Document Parsing
pdfplumber==0.10.3
python-docx==1.1.0
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Development
pytest==7.4.3
httpx==0.25.2
